requirements.txt
PyPDF2==3.0.1
pdfplumber==0.10.3
Pillow==10.1.0
pytesseract==0.3.10
pymupdf==1.24.7
